modificador de mensajes de chat y titulos
