+++ AppChat_utn_nivelAvanzado/AppChat_utn_nivelAvanzado/client/controllers/login.py
@@ -1,6 +1,7 @@
 from PySide2.QtWidgets import QWidget
 from views.login import LoginForm
-
+from pathlib import Path
+from PySide2.QtGui import QIcon
 from pys2_msgboxes import msgboxes
 
 class LoginWindow(QWidget, LoginForm):
@@ -13,10 +14,14 @@ class LoginWindow(QWidget, LoginForm):
 
     def join_to_chat(self):
         username = self.usernameLineEdit.text()
-
+        token = ""
         if username != '':
             from controllers.chat import ChatWindow
-            self.chat_window = ChatWindow(username)
+            ruta_icono = Path('.', 'assets', 'logo.png')
+            self.chat_window = ChatWindow(username, token)
+            self.chat_window.setWindowTitle('UTN CHAT')
+            self.chat_window.setWindowIcon(QIcon(str(ruta_icono)))
+            self.chat_window.setFixedSize(877, 538)
             self.chat_window.show()
             self.close()
         else:
