+++ AppChat_utn_nivelAvanzado/AppChat_utn_nivelAvanzado/requirements.txt
@@ -1,0 +1,3 @@
+pys2-msgboxes==0.0.2
+PySide2==5.15.2
+shiboken2==5.15.2
