+++ AppChat_utn_nivelAvanzado/AppChat_utn_nivelAvanzado/utn_chat_info.log
@@ -1,0 +1,11 @@
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:11:16', 'Iniciando Server...', '[STARTING]', '192.168.0.190', 55555)
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:11:16', '[SERVER] Corriendo OK!')
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:11:16', '[LISTENING] Esperado Conexiones en IP:', '192.168.0.190', 'puerto:', 55555)
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:11:49', 'Client Hexadecimal token :', 'e163')
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:11:49', 'user id: ', 'fede', 'Conectado OK')
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:11:49', '[CONEXIONES ACTIVAS]', 1)
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:12:01', 'fede', 'Desconectado')
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:13:44', 'Client Hexadecimal token :', 'e163')
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:13:44', 'user id: ', 'FEDE', 'Conectado OK')
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:13:44', '[CONEXIONES ACTIVAS]', 1)
+INFO:root:UTN Chat App log: ('26.Nov 2021 19:13:51', 'FEDE', 'Desconectado')
